优化音乐播放功能
