+++ app/src/main/res/layout/fragment_music.xml
@@ -23,7 +23,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="31dp"
-        android:text="Hello World!"
+        android:text="@string/to"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="@+id/seekBar"
@@ -47,7 +47,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
-        android:text="@string/textview"
+        android:text="@string/musicname"
         app:layout_constraintEnd_toEndOf="@+id/textView_mn"
         app:layout_constraintHorizontal_bias="0.156"
         app:layout_constraintStart_toStartOf="@+id/textView_mn"
@@ -91,7 +91,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:text="@string/stop"
+        android:text="@string/restore"
         app:layout_constraintEnd_toEndOf="@+id/button4"
         app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="@+id/button4"
