+++ app/src/main/res/values/strings.xml
@@ -7,22 +7,25 @@
     <string name="title_chat">聊天</string>
     <string name="timer">00:00:00</string>
     <string name="start">start</string>
-    <string name="stop">stop</string>
+    <string name="restore">继续</string>
+    <string name="stop">停止</string>
     <string name="restart">restart</string>
     <string name="textview">TextView</string>
     <string name="textview1">TextView</string>
     <string name="textview2">TextView</string>
     <string name="textview3">TextView</string>
     <string name="todo">TODO</string>
-    <string name="score">Score:</string>
-    <string name="reset">Reset</string>
+    <string name="score">分数:</string>
+    <string name="reset">重置</string>
     <string name="a">A*</string>
-    <string name="play">play</string>
-    <string name="prev">prev</string>
-    <string name="next">next</string>
-    <string name="pause">pause</string>
-    <string name="notification">notification</string>
+    <string name="play">播放</string>
+    <string name="prev">上一首</string>
+    <string name="next">下一首</string>
+    <string name="pause">暂停</string>
+    <string name="notification">通知</string>
     <string name="chat">请输入聊天内容</string>
     <string name="send">发送</string>
     <string name="text">你好!</string>
+    <string name="musicname">音乐名</string>
+    <string name="to">播放至</string>
 </resources>
